day 10 accomplished
There is a loss of precision leading to inaccuracies in densely populated
portions of the map.  I think this is because I computed only integer
based slopes, so because -- as the problem states-- that while the grid
is indexed as boxes, the asteroid is only in the very center.  So it is
possible that there is the possibility that the laser may be able to hit
an asteroid that appears to be behind another at low precision, but
at high precision it may superceded in the destruction order.

Plan to reimplement with high precision